requirements.txt
numpy==1.24.4
torch==2.0.1
transformers==4.31.0
spacy==3.5.3
tqdm
PyMuPDF
python-docx
pytesseract
pillow